+++ Recording Student Achievements/Recording Student Achievements/Recording Student Achievements.csproj
@@ -55,19 +55,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CustomClasses\NSNSearch.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="CustomClasses\GeekIt.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="CustomClasses\AddAssessment.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="CustomClasses\NumericTextBox.cs">
-      <SubType>Component</SubType>
-    </Compile>
-    <Compile Include="CustomClasses\CustomRectangle.cs" />
+    <Compile Include="CustomRectangle.cs" />
     <Compile Include="IndividualReport.cs">
       <SubType>Form</SubType>
     </Compile>
