--- Recording%20Student%20Achievements/Recording%20Student%20Achievements/CustomClasses/NumericTextBox.cs
@@ -1,86 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace Recording_Student_Achievements
-{
-    public class NumericTextBox : TextBox
-    {
-        bool allowSpace = false;
-
-        // Restricts the entry of characters to digits (including hex), the negative sign,
-        // the decimal point, and editing keystrokes (backspace).
-        protected override void OnKeyPress(KeyPressEventArgs e)
-        {
-            base.OnKeyPress(e);
-
-            NumberFormatInfo numberFormatInfo = System.Globalization.CultureInfo.CurrentCulture.NumberFormat;
-            string decimalSeparator = numberFormatInfo.NumberDecimalSeparator;
-            string groupSeparator = numberFormatInfo.NumberGroupSeparator;
-            string negativeSign = numberFormatInfo.NegativeSign;
-
-            string keyInput = e.KeyChar.ToString();
-
-            if (Char.IsDigit(e.KeyChar))
-            {
-                // Digits are OK
-            }
-            else if (keyInput.Equals(decimalSeparator) || keyInput.Equals(groupSeparator) ||
-             keyInput.Equals(negativeSign))
-            {
-                // Decimal separator is OK
-            }
-            else if (e.KeyChar == '\b')
-            {
-                // Backspace key is OK
-            }
-            //    else if ((ModifierKeys & (Keys.Control | Keys.Alt)) != 0)
-            //    {
-            //     // Let the edit control handle control and alt key combinations
-            //    }
-            else if (this.allowSpace && e.KeyChar == ' ')
-            {
-
-            }
-            else
-            {
-                // Swallow this invalid key and beep
-                e.Handled = true;
-                //    MessageBeep();
-            }
-        }
-
-        public int IntValue
-        {
-            get
-            {
-                return Int32.Parse(this.Text);
-            }
-        }
-
-        public decimal DecimalValue
-        {
-            get
-            {
-                return Decimal.Parse(this.Text);
-            }
-        }
-
-        public bool AllowSpace
-        {
-            set
-            {
-                this.allowSpace = value;
-            }
-
-            get
-            {
-                return this.allowSpace;
-            }
-        }
-    }
-}
